requirements.txt
# Python 3.12
# ============================================================
# PyTorch 生态（CUDA 12.8）
# CPU 环境请改为：pip install torch torchvision torchaudio
# ============================================================
--extra-index-url https://download.pytorch.org/whl/cu128
torch==2.10.0+cu128
torchvision==0.25.0+cu128
torchaudio==2.10.0+cu128

# ============================================================
# JAX 生态（CUDA 12.x）
# CPU 环境请改为：pip install "jax[cpu]"
# ============================================================
jax[cuda12]>=0.4.35
flax>=0.10.0
optax>=0.2.4
orbax-checkpoint>=0.10.0

# ============================================================
# 数据集
# ============================================================
huggingface_hub>=0.27.0
datasets>=3.3.0
pyarrow>=18.0.0
Pillow>=11.0.0

# ============================================================
# TensorBoard
# ============================================================
tensorboard>=2.19.0
tensorboardX>=2.6.2

# ============================================================
# ONNX 导出与推理
# ============================================================
onnx>=1.17.0
onnxruntime-gpu>=1.21.0

# ============================================================
# 模型分析（参数量、FLOPs）
# ============================================================
thop>=0.1.1

# ============================================================
# 工具
# ============================================================
numpy>=2.0.0
tqdm>=4.67.0
